refactor: Add SelfInfo component
