+++ client/modules/main/chatPanel/chat/Chat.less
@@ -605,6 +605,7 @@
             font-size: 14px;
             margin-left: 6px;
             margin-top: 4px;
+            color: #444;
         }
         &:last-child {
             margin-bottom: 0px;
